requirements.txt
Flask
azure-ai-inference
gunicorn